+++ src/runtime/@types/common.d.ts
@@ -1,0 +1,7 @@
+import {TabsController} from '..';
+
+declare global {
+    interface Window {
+        diplodocTabs: TabsController;
+    }
+}
